Refactor XREffect and other logic out into their own modules
