+++ .gitignore
@@ -1,2 +1,3 @@
 /vulkan/build/
 /build/
+/gnome/setup-45.sh
