+++ gnome/breezydesktop@org.xronlinux/globals.js
@@ -1,0 +1,5 @@
+const Globals = {
+    ipc_file:       null,   // Gio.File instance, file exists
+    extension_dir:  null    // string path
+}
+export default Globals;
