+++ gnome/breezydesktop@org.xronlinux/math.js
@@ -1,0 +1,3 @@
+export function degreeToRadian(degree) {
+    return degree * Math.PI / 180;
+}
