+++ gnome/breezydesktop@org.xronlinux/time.js
@@ -1,0 +1,3 @@
+export function getEpochSec() {
+    return Math.floor(Date.now() / 1000);
+}
